// apps/studio/src/front/features/models/pending-requests/resource.ts
export enum RSC {
  MODELS_REQUESTS_SUMMARY_MESSAGE = "models.requests.summary.message",
  MODELS_REQUESTS_UNKNOWN_TEXT = "models.requests.unknown.text",
  MODELS_REQUESTS_REFRESH_BUTTON = "models.requests.refresh.button",
  MODELS_REQUESTS_CLEAR_BUTTON = "models.requests.clear.button",
  MODELS_REQUESTS_CONFIRM_TITLE = "models.requests.confirm.title",
  MODELS_REQUESTS_CONFIRM_BUTTON = "models.requests.confirm.button",
  MODELS_REQUESTS_WARNING_MESSAGE = "models.requests.warning.message",
  MODELS_REQUESTS_CANCEL_BUTTON = "models.requests.cancel.button",
  MODELS_REQUESTS_STAGE_MESSAGE = "models.requests.stage.message",
  MODELS_REQUESTS_OWNERS_MESSAGE = "models.requests.owners.message",
  MODELS_REQUESTS_TIME_MESSAGE = "models.requests.time.message",
  MODELS_REQUESTS_CLEARED_MESSAGE = "models.requests.cleared.message",
  MODELS_REQUESTS_RECOVERY_BUTTON = "models.requests.recovery.button",
  MODELS_REQUESTS_ERROR_MESSAGE = "models.requests.error.message",
  MODELS_REQUESTS_HISTORY_MESSAGE = "models.requests.history.message",
  MODELS_REQUESTS_INSPECTION_ERROR_MESSAGE = "models.requests.inspection.error.message",
}